// src/app/interfaces/interface-minuta.ts
export interface IMinuta {
  filial: string;
  serie: string;
  documento: string;
  cliente: string;
  nomeCliente: string;
  cidade: string;
  estado: string;
  valor: string;
  especie: string;
  peso: string;
  volume: string;
  transportadora: string;
  assinatura: string;
  emissao: string;
  motorista: string;
  placa: string;
  cpf: string;
  conferente: string;
}

    